+++ aws/rust-runtime/aws-config/src/imds/client.rs
@@ -11,8 +11,7 @@ use crate::imds::client::error::{BuildError, ImdsError, InnerImdsError, InvalidE
 use crate::imds::client::token::TokenRuntimePlugin;
 use crate::provider_config::ProviderConfig;
 use crate::PKG_VERSION;
-use aws_http::user_agent::{ApiMetadata, AwsUserAgent};
-use aws_runtime::user_agent::UserAgentInterceptor;
+use aws_runtime::user_agent::{ApiMetadata, AwsUserAgent, UserAgentInterceptor};
 use aws_smithy_runtime::client::orchestrator::operation::Operation;
 use aws_smithy_runtime::client::retries::strategy::StandardRetryStrategy;
 use aws_smithy_runtime_api::box_error::BoxError;
