+++ aws/rust-runtime/aws-http/Cargo.toml
@@ -1,25 +1,14 @@
 [package]
 name = "aws-http"
-version = "0.0.0-smithy-rs-head"
+version = "0.60.5"
 authors = ["AWS Rust SDK Team <aws-sdk-rust@amazon.com>", "Russell Cohen <rcoh@amazon.com>"]
-description = "HTTP specific AWS SDK behaviors."
+description = "This crate is no longer used by the AWS SDK and is deprecated."
 edition = "2021"
 license = "Apache-2.0"
 repository = "https://github.com/smithy-lang/smithy-rs"
 
 [dependencies]
-aws-smithy-runtime-api = { path = "../../../rust-runtime/aws-smithy-runtime-api", features = ["client"] }
-aws-smithy-types = { path = "../../../rust-runtime/aws-smithy-types", features = ["http-body-0-4-x"] }
-aws-types = { path = "../aws-types" }
-bytes = "1.1"
-http = "0.2.3"
-http-body = "0.4.5"
-tracing = "0.1"
-pin-project-lite = "0.2.9"
-
-[dev-dependencies]
-bytes-utils = "0.1.2"
-tokio = { version = "1.23.1", features = ["macros", "rt", "time"] }
+aws-runtime = { path = "../aws-runtime", features = ["http-02x"] }
 
 [package.metadata.docs.rs]
 all-features = true
