+++ aws/rust-runtime/aws-http/README.md
@@ -1,9 +1,6 @@
 # aws-http
 
-This crate provides middleware for AWS SDKs using HTTP including:
-* Generalized retry policy
-* Middleware for setting `User-Agent` headers based on runtime configuration
-* Credential loading async middleware
+This crate is no longer used by the AWS SDK and is deprecated.
 
 <!-- anchor_start:footer -->
 This crate is part of the [AWS SDK for Rust](https://awslabs.github.io/aws-sdk-rust/) and the [smithy-rs](https://github.com/smithy-lang/smithy-rs) code generator. In most cases, it should not be used directly.
