+++ aws/rust-runtime/aws-http/external-types.toml
@@ -1,11 +1,3 @@
 allowed_external_types = [
-    "aws_smithy_runtime_api::http::headers::Headers",
-    "aws_smithy_types::body::Error",
-    "aws_smithy_types::config_bag::storable::Storable",
-    "aws_smithy_types::config_bag::storable::StoreReplace",
-    "aws_smithy_types::error::metadata::Builder",
-    "aws_types::app_name::AppName",
-    "aws_types::os_shim_internal::Env",
-    "bytes::bytes::Bytes",
-    "http_body::Body",
+    "aws_runtime::*",
 ]
